Don't use scene name from sceneLoaded args.
There seems to be more to `sceneLoaded` than I thought. The false
knight room first loads as `Crossroads_10` then immediately after as
`Crossroads_10_boss` but all the grass is located in `Crossroads_10`
still. The end result is that the status is wrong until you hit some
grass.

Looks like the global is always right, with a little bit of a delay.
Easy enough.