Add grass aliases.
Who knew this would be needed... Skurry found a very odd set of grass
in the abyss. I have no idea why or what triggers this, but there's a
ledge that sometimes has 3 grass objects of particular names, or
sometimes there's 2 grass objects of different particular names.

The extra grass object is currently ignored by the mod. I plan to fix
this issue in a separate release.

But the other 2 share the exact same location so I believe that
considering them to just be duplicates of eachother is the right move.

I have not confirmed that I've gotten _all_ possible aliases that might
appear on that ledge, and have only gone through and tested it a few
times. But it seems to be working.