Add the quantum grass to the list.
This grass is called the "quantum grass" because sometimes its there
and sometimes its not. For awhile no one anywhere in the world knew
why. But now the knowledge has entered the hands of mortals.

The grass is loaded if you enter the scene without wings. The grass is
not loaded if you enter the scene with wings.

Mystery solved! Adding the grass to the list now. Players will need to
cut it before they get rings or their run will be dead. A surprising
fact for sure, but I think that's fine.