Revert "Fade-in grass count with geo count."
This reverts commit dd1d88463ac656d58334a3a5be766c0f97a7e957.

This wasn't doing what I thought it was doing. It really only helps
with the fade-in at the beginning of the game. Fade-in during screen
transitions I got for free with how I placed the counter.

I'd like to fade-in better at the beginning of the game. This isn't how
to accomplish that though, because tying it to the geo counter has a
bunch of problems.