Disable compass by default. Show it on load.
Grassy Knight has all the features a speedrunner would want right now,
so I'm preparing a speedrunner-centric release.

This commit makes it so the compass is completely disabled by default
and you have to go into the settings to enable it. But once you've got
it enabled, it will appear by default on load.

This is done with the expectation that a speedrunner will only enable
the compass during practice. And as the mod is much more battle-tested
now, the need for a compass has lessened.