Add zip-file dependencies to Makefile. Fix clean.
This has been bugging me for a bit... had a moment so I figured I'd
patch this.

Clean would fail if any of the files-to-delete weren't there. And most
of the targets under zip-file/ would fail if zip-file/ hadn't been
created yet. The way Makefile ordered the dependencies meant that if
just specified `GrassyKnight.zip` as the target it'd just work though,
so all this is pretty pedantic. 🤷‍♀️